Update pypi metadata to reference -infra
Most of this was existing cookiecutter defaults used for OpenStack
projects. Additionally, we are adding the alpha tag since we are still
early in the release process.  Over the next few releases we'll bump
this up to beta then eventually stable.

Also, http://docs.openstack.org/infra/system-config/grafyaml.html
currently does not exist. I will be adding a follow-up patch for
system-config.

Change-Id: Idd601c3179cbec87441713f5317d51b0ebe134b1
Signed-off-by: Paul Belanger <pabelanger@redhat.com>