Merge json support (#18)
* Add import of json files

This simply takes any json files present and loads them into Grafana
directly.  The idea is that you can edit the dashboards using the
inbuilt editor, then copy the dashboard JSON and keep it externally
version controlled.  No parsing or validation is done on the JSON
files; we are assuming they have not been hand-modified from what
Grafana generates.

Change-Id: I38695aed2404f8b7fc350d949b7a9212498c35cb

* Build docker images

This builds and publishes docker images to opendevorg/grafyaml
so that users can run this with "docker run ...".

Also, appease the pep8 gods.

Change-Id: Ia30438953e5c73a01ae0cd32afb660442e0e6e39

* Remove unused files

* Trigger CI

* Run lint

Co-authored-by: Ian Wienand <iwienand@redhat.com>
Co-authored-by: James E. Blair <jim@acmegating.com>
Co-authored-by: Zuul <zuul@review.opendev.org>